Correctly setup context
